check for _LARGEFILE64_SOURCE instead of __USE_LARGEFILE64
glibc doesn't define getdents() and requires use of the raw syscall. As
such lib9/dirread.c needed to decide whether to use SYS_getdents or
SYS_getdents64 and was checking if __USE_LARGEFILE64 was defined in
order to do so. musl does not define __USE_LARGEFILE64 so the wrong
syscall was being used. musl does however define _LARGEFILE64_SOURCE,
the macro that glibc checks in order to define __USE_LARGEFILE64.